# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SyncRequest {
  /** All completed topic slugs */
  completedSlugs: string[];
  /** All bookmarked topic slugs */
  bookmarkedSlugs: string[];
  /** All earned achievement IDs */
  achievementIds: string[];
  /** @minimum 0 */
  currentStreak: number;
  /** @minimum 0 */
  bestStreak: number;
  lastActivityDate?: string | null;
  displayName?: string | null;
  email: string;
}

export interface SyncResponse {
  completedSlugs: string[];
  bookmarkedSlugs: string[];
  achievementIds: string[];
  currentStreak: number;
  bestStreak: number;
  lastActivityDate?: string | null;
  syncedAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  teamName?: string | null;
  completedCount: number;
  currentStreak: number;
  bestStreak: number;
  isCurrentUser: boolean;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  currentUserRank?: number | null;
}

export interface UserProfile {
  id: string;
  email: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  teamName?: string | null;
  completedCount: number;
  currentStreak: number;
  bestStreak: number;
  createdAt: string;
}

export interface UpdateProfileRequest {
  displayName?: string | null;
  teamName?: string | null;
}

export type GetLeaderboardParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};