# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  teamName?: string | null;
  completedCount: number;
  currentStreak: number;
  bestStreak: number;
  isCurrentUser: boolean;
}